Refactored the code for when trip session pages, dividing components and pages up
