+++ frontend-android/app/src/main/java/com/example/wetravel/models/DataModels.kt
@@ -1,0 +1,21 @@
+package com.example.wetravel.models
+
+data class Destination(
+    val name: String,
+    val address: String,
+    val rating: String,
+    val reviewCount: Int,
+    val imageResId: Int,
+    val numOfVotes: Int,
+    val voted: Boolean
+)
+
+
+
+
+
+
+
+
+
+
